fix: super cmd with permissions
