+++ apps/artemis-web/components/401-hoc/401-hoc.tsx
@@ -2,10 +2,19 @@ import { useJWT } from '../../utils/hooks/use-jwt';
 import DefaultErrorPage from 'next/error';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
-import React from 'react';
+import React, { useState } from 'react';
+import { GlobalHotKeys } from 'react-hotkeys';
+import { commands, handlers, keyMap } from '../../utils/power-actions';
+import { OfflineBoltOutlined } from '@material-ui/icons';
+import CommandLineModal from 'react-super-cmd';
+
 
 const AuthHOC = (WrappedComponent, ACL = []) => {
   const Wrapped = (props) => {
+    const [cmdLineModal, setCmdLineModal] = useState(false);
+    function toggleIsOpen() {
+      setCmdLineModal((previousState) => !previousState);
+    }
     const [user, loading] = useJWT();
     const router = useRouter();
     if (!user && !loading) {
@@ -25,7 +34,20 @@ const AuthHOC = (WrappedComponent, ACL = []) => {
         </>
       );
     } else if (!loading) {
-      return <WrappedComponent {...props} user={user}></WrappedComponent>;
+      return (
+        <GlobalHotKeys keyMap={keyMap} handlers={handlers(toggleIsOpen)}>
+          <WrappedComponent {...props} user={user}></WrappedComponent>
+          <CommandLineModal
+            commands={commands}
+            isOpen={cmdLineModal}
+            toggleIsModalOpen={toggleIsOpen}
+            title={'Navigate with the keyboard'}
+            logo={<OfflineBoltOutlined />}
+            noOptionsText="No commands found. Try a different search term."
+          />
+        </GlobalHotKeys>
+      )
+        ;
     } else {
       return <> </>;
     }
