+++ apps/artemis-web/components/layout/layout.tsx
@@ -7,7 +7,6 @@ export default class Layout extends React.Component<any> {
   render() {
     const { children } = this.props;
     const props = this.props;
-    // const Footer = dynamic(() => import('../footer/footer'));
     const Header = dynamic(() => import('../header/header'));
 
     return (
